+++ lib/src/dart_formatter.dart
@@ -90,6 +90,7 @@ class DartFormatter {
     var featureSet = FeatureSet.fromEnableFlags2(
       sdkLanguageVersion: Version(2, 19, 0),
       flags: [
+        'records',
         'unnamed-libraries',
       ],
     );